# openapi/swagger.yaml
openapi: 3.1.0
info:
  title: JARVIS Swap API
  version: 1.0.0-rc.15
  description: Server-side Sui network, wallet, price, pool and swap quote boundaries for JARVIS Swap.
servers:
  - url: /api/v1
paths:
...
  /network/status:
    get:
      operationId: getSuiNetworkStatus
      responses:
        '200': { description: Sui checkpoint, gas reference price and RPC latency }
        '503': { description: Sui RPC unavailable }
...
components:
  securitySchemes:
    cronBearer:
      type: http
      scheme: bearer
      bearerFormat: CRON_SECRET
...